backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.1
cryptography==42.0.2
django-cors-headers==4.3.1
django-filter==24.1
Pillow==10.2.0
python-dateutil==2.8.2
openpyxl==3.1.2
